// src/model/sharepoint/SearchResultWithAnalytics.ts
export interface SearchResultWithAnalytics {
    Title: string;
    SiteTitle: string;
    SiteName: string;
    SiteId: string;
    WebId: string;
    ListId: string;
    DriveId: string;
    ListItemId: string;
    UniqueId: string;
    NormUniqueId: string;
    SPWebUrl: string;
    ViewsLifeTime: number;
    ViewsLifeTimeUniqueUsers: number;
    ViewsLast1Day: number;
    ViewsLast1DayUniqueUsers: number;
    ViewsLast2Days: number;
    ViewsLast2DaysUniqueUsers: number;
    ViewsLast3Days: number;
    ViewsLast3DaysUniqueUsers: number;
    ViewsLast4Days: number;
    ViewsLast4DaysUniqueUsers: number;
    ViewsLast5Days: number;
    ViewsLast5DaysUniqueUsers: number;
    ViewsLast6Days: number;
    ViewsLast6DaysUniqueUsers: number;
    ViewsLast7Days: number;
    ViewsLast7DaysUniqueUsers: number;
    ViewsLastMonths1: number;
    ViewsLastMonths1Unique: number;
    ViewsLastMonths2: number;
    ViewsLastMonths2Unique: number;
    ViewsLastMonths3: number;
    ViewsLastMonths3Unique: number;
    SPTranslatedLanguages: string;
    SPTranslationLanguage: string;
    CommentCount: number;
    LikeCount: number;
    Path: string;
    LastModifiedTime: string;
    Author: string;
    FileExtension: string;
}